wm: move classifer to workload_service
