+++ ydb/core/kqp/proxy_service/ut/ya.make
@@ -11,7 +11,6 @@ ENDIF()
 
 SRCS(
     kqp_proxy_ut.cpp
-    kqp_query_classifier_ut.cpp
     kqp_script_executions_ut.cpp
 )
