+++ ydb/core/kqp/proxy_service/ya.make
@@ -4,7 +4,6 @@ SRCS(
     kqp_proxy_databases_cache.cpp
     kqp_proxy_peer_stats_calculator.cpp
     kqp_proxy_service.cpp
-    kqp_query_classifier.cpp
     kqp_query_text_cache_service.cpp
     kqp_script_executions.cpp
     kqp_session_info.cpp
