+++ ydb/core/kqp/workload_service/ut/ya.make
@@ -10,6 +10,7 @@ ELSE()
 ENDIF()
 
 SRCS(
+    kqp_query_classifier_match_ut.cpp
     kqp_query_classifier_ut.cpp
     kqp_workload_service_actors_ut.cpp
     kqp_workload_service_query_sessions_ut.cpp
