+++ ydb/core/kqp/workload_service/ya.make
@@ -1,6 +1,7 @@
 LIBRARY()
 
 SRCS(
+    kqp_query_classifier.cpp
     kqp_workload_service.cpp
 )
 
@@ -9,11 +10,19 @@ PEERDIR(
 
     ydb/core/fq/libs/compute/common
 
+    ydb/core/kqp/common
+    ydb/core/kqp/gateway/behaviour/resource_pool_classifier
+    ydb/core/kqp/query_data
     ydb/core/kqp/workload_service/actors
 
     ydb/core/mind
 
+    ydb/core/resource_pools
+
     ydb/library/actors/interconnect
+    ydb/library/aclib
+
+    ydb/public/api/protos
 )
 
 YQL_LAST_ABI_VERSION()
